Implement run.download_artifact() (#828)
* Enable clientside artifact store

* Add test

* Add warning TODO

* Downgrade TODO to NOTE

* Capture env var closer to when actually used

* Improve print messages

* Fix test cleanup

* Impl download_artifact()

* No need to complicate

* Move download_artifact() to sep PR

* Impl download_artifact()

* No need to complicate

* Fix comment

* Add path-only error test

* Revert "No need to complicate"

This reverts commit 3153ba67a62b5658845a6a92bf0af8a28b1bd1ce.

* Don't read into memory

* Clarify TODOs

* Improve test

* There's no tempdir anymore

* Test object download

* Another check

* Increase default chunk size to 32MB


* Use copyfile

* Update docstring

+++ client/verta/verta/dataset/_dataset.py
@@ -159,7 +159,7 @@ class _Dataset(blob.Blob):
                 os.rename(tempf.name, download_to_path)
             except Exception as e:
                 # delete partially-downloaded file
-                if tempf is not None:
+                if tempf is not None and os.path.isfile(tempf.name):
                     os.remove(tempf.name)
                 raise e
             else:
